pyproject.toml
[tool.poetry]
name = "noise-cli"
version = "0.1.0"
description = ""
authors = ["Max Fan <root@max.fan>"]

[tool.poetry.dependencies]
python = "^3.9"
noiseprotocol = "^0.3.1"
typer-cli = "^0.0.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
